+++ src/twister/models/Plateau.java
@@ -40,4 +40,7 @@ public class Plateau {
 		}
 		return liste;
 	}
+	
+	
+	
 }
